build: Successfully build/compile on MS Windows

+++ upnp/src/genlib/net/http/httpreadwrite.c
@@ -60,7 +60,7 @@
 //#include <string.h>
 
 #ifdef _WIN32
-//#include <malloc.h>
+#include <malloc.h>
 #define fseeko fseek
 #if defined(_MSC_VER) && _MSC_VER < 1900
 #define snprintf _snprintf
@@ -1737,7 +1737,7 @@ int http_MakeMessage(membuffer *buf,
                 } else if (c == 'd') {
                         /* integer */
                         num = (size_t)va_arg(argp, int);
-                        rc = snprintf(tempbuf, sizeof(tempbuf), "%" PRIzu, num);
+                        rc = snprintf(tempbuf, sizeof(tempbuf), "%" PRIzu, (unsigned long)num);
                         if (rc < 0 || (unsigned int)rc >= sizeof(tempbuf) ||
                                 membuffer_append(buf, tempbuf, strlen(tempbuf)))
                                 goto error_handler;
